+++ packages/yak_runner/lib/src/delegates/test/factory/nullary.dart
@@ -1,8 +1,8 @@
 import '../delegate/all.dart';
 
-/// a `mixin` that allow to generate `YakRunnerTestDelegate`
+/// a `mixin` that allow to generate `RunnerTestDelegate`
 /// of the same type `T` of the the class calling the function
-mixin YakRunnerTestFactoryDelegate {
-  /// `call` returns a `YakRunnerTestDelegate` of type `T`
-  YakRunnerTestDelegate<T> call<T>();
+mixin RunnerTestFactoryDelegate {
+  /// `call` returns a `RunnerTestDelegate` of type `T`
+  RunnerTestDelegate<T> call<T>();
 }