+++ packages/yak_runner/lib/src/extensions/custom/iterable_contains.dart
@@ -1,8 +1,8 @@
 import '../../all.dart';
 
 /// an extension on `Iterable` to easily access the result
-extension YakRunnerIterableContainsExtension<T> on Iterable<T> {
-  /// returns a `YakRunnerArg` from an Iterable
+extension RunnerIterableContainsExtension<T> on Iterable<T> {
+  /// returns a `RunnerArg` from an Iterable
   Result<T> containsResult(T element) {
     final runner = IterableContains(this);
     return runner(element);