+++ packages/yak_runner/lib/src/extensions/failure/of_type.dart
@@ -1,0 +1,7 @@
+import '../../../yak_runner.dart';
+
+/// cast a `Failure<S>` to a `Failure<T`
+extension FailureToFailureOfTypeX<T, S> on Failure<S> {
+  /// matches the typedef `FailureOfType<T>`
+  Failure<T> ofType<T>() => Failure<T>(e, s);
+}