+++ packages/yak_runner/lib/src/mixin/test/nullary.dart
@@ -1,10 +1,9 @@
 import '../../all.dart';
 
-/// a `mixin` on `YakRunnerBase`
-mixin YakRunnerTestMixin<T> on YakRunnerBase {
-  /// is a `YakRunnerTestDelegate Function(YakRunnerTestFactoryDelegate)`
-  /// type `T` is passed by the `Object` to the test to `YakRunnerTestDelegate`
-  YakRunnerTestDelegate<T> buildTest(
-          YakRunnerTestFactoryDelegate testFactory) =>
+/// a `mixin` on `RunnerBase`
+mixin RunnerTestMixin<T> on RunnerBase {
+  /// is a `RunnerTestDelegate Function(RunnerTestFactoryDelegate)`
+  /// type `T` is passed by the `Object` to the test to `RunnerTestDelegate`
+  RunnerTestDelegate<T> buildTest(RunnerTestFactoryDelegate testFactory) =>
       testFactory<T>();
 }