+++ packages/yak_runner/test/mixin/test/test_factory/unary.dart
@@ -2,8 +2,8 @@ import 'package:yak_runner/yak_runner.dart';
 
 import '../test_delegate/unary.dart';
 
-class FakeUnaryTestFactory implements YakRunnerArgTestFactoryDelegate {
+class FakeUnaryTestFactory implements UnaryRunnerTestFactoryDelegate {
   const FakeUnaryTestFactory();
   @override
-  YakRunnerArgTestDelegate<T, S> call<T, S>() => const FakeUnaryTest();
+  UnaryRunnerTestDelegate<T, S> call<T, S>() => const FakeUnaryTest();
 }