+++ packages/yak_runner/test/mocks/delegates/function/nullary.dart
@@ -1,7 +1,7 @@
 import 'package:stub/stub.dart';
 import 'package:yak_runner/yak_runner.dart';
 
-class MockDelegate<T> extends Stub<T> implements Delegate<T> {
+class MockDelegate<T> extends Stub<T> implements NullaryDelegate<T> {
   @override
   T call() => stub();
 }