+++ packages/yak_runner/test/mocks/delegates/test/delegate/nullary.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 import 'package:stub/stub.dart';
 import 'package:yak_runner/yak_runner.dart';
 
-class MockYakRunnerTestDelegate<T> extends Stub<void>
-    implements YakRunnerTestDelegate<T> {
+class MockRunnerTestDelegate<T> extends Stub<void>
+    implements RunnerTestDelegate<T> {
   @override
   void call(FutureOr<T> result) => stub();
 }