+++ packages/yak_runner/test/mocks/delegates/test/delegate/unary.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 import 'package:stub/stub.dart';
 import 'package:yak_runner/yak_runner.dart';
 
-class MockYakRunnerArgTestDelegate<T, S> extends Stub<void>
-    implements YakRunnerArgTestDelegate<T, S> {
+class MockUnaryRunnerTestDelegate<T, S> extends Stub<void>
+    implements UnaryRunnerTestDelegate<T, S> {
   @override
   void call(FutureOr<T> result, FutureOr<S> arg) => stub();
 }